Find a small trick for CUDA
When executing heterogeneouse codes ( c++ plus CUDA C), the first CUDA
sentence consume much more time than others. This is because CUDA needs
to be initialized and the run time library needs to be linked.
Addionally, the latencies between host and device should be taken.
Solution: add an cudaFree(0) in the initialization phase before
excecuation.